packages/poly_ren/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "poly-ren"
version = "1.0.0"
description = "Renpy-specific text extraction and merging tools for translation workflows"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Enro Team"}
]
keywords = ["renpy", "translation", "extraction", "visual-novel"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "pyyaml>=6.0",
    # For monorepo: imports from src.poly_trans
    # For PyPI release: "poly-trans>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "mypy>=1.0",
]

[tool.setuptools]
packages = ["poly_ren"]
